requirements.txt
aiogram
python-dotenv
requests
gigachain
gigachain-cli
gigachain_community